+++ changelog.md
@@ -22,6 +22,7 @@
 - Add K3S Framework API Server into the Container
 - Add API to get Kubeconfig
 - Add Kubernetes server check to scheduler the agent only if the server is running
+- Add automatic Kubernetes dashboard deploy to the bootstrap
 
 ## v0.0.4